+++ db/migrate/20260101000021_add_body_format_to_bible270_comments.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class AddBodyFormatToBible270Comments < ActiveRecord::Migration[7.0]
+  def up
+    add_column :bible270_comments, :body_format, :string, null: false, default: 'plain'
+  end
+
+  def down
+    remove_column :bible270_comments, :body_format
+  end
+end